+++ lib/generators/cms/attribute/templates/modules/date_attribute.rb
@@ -3,7 +3,7 @@ module Cms
     # This is a date attribute concern. It should be included via +include
     # Cms::Attributes::<%= class_name %>+ in all CMS models that use this attribute.
     module <%= class_name %>
-      def <%= file_name %>
+      def <%= method_name %>
         self[:<%= file_name %>]
       end
     end