+++ lib/generators/cms/attribute/templates/modules/linklist_attribute.rb
@@ -3,17 +3,17 @@ module Cms
     # This is a linklist attribute concern. It should be included via +include
     # Cms::Attributes::<%= class_name %>+ in all CMS models that use this attribute.
     module <%= class_name %>
-      def <%= file_name %>
+      def <%= method_name %>
         self[:<%= file_name %>] || RailsConnector::LinkList.new(nil)
       end
 
-      def <%= file_name %>?
-        <%= file_name %>.present?
+      def <%= method_name %>?
+        <%= method_name %>.present?
       end
 
       <%- if max_size == 1 -%>
-      def first_<%= file_name %>
-        <%= file_name %>.destination_objects.first
+      def first_<%= method_name %>
+        <%= method_name %>.destination_objects.first
       end
       <%- end -%>
     end