+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.0.0
   * Renamed gem from `ice_kickstarter` to `infopark_kickstarter`.
+  * Added attribute generator option to set the name of getter method. (Thanks @cocodercoder)
   * Added support for integer and float attribute types.
   * Bugfix: Newrelic generator did not insert the correct website name in the deploy files. It also
     does not depend on the kickstart generator anymore.